pyproject.toml
[tool.poetry]
name = "tater"
version = "0.2.0"
description = "A Discord bot that uses Ollama, Redis, SFTPGo, and Discord.py for conversation continuity and character-based responses."
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.2.2"           # Discord library for the bot
redis = "^4.5.3"                # For managing conversation history
ollama = "^0.1.0"               # Python wrapper for Ollama's API
python-dotenv = "^1.0.0"        # For loading environment variables from .env file
aiohttp = "^3.8.4"              # For async HTTP requests (SFTPGo API)
requests = "^2.31.0"            # For synchronous HTTP requests (YouTube API)

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"               # For testing if needed

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"